Delete verification Country League Player

+++ backofficeApp/app/Http/Controllers/CountryController.php
@@ -4,6 +4,9 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Country;
+use App\Models\Event;
+use App\Models\Team;
+use App\Models\LeagueCountry;
 use Illuminate\Support\Facades\Validator;
 use \Illuminate\Database\QueryException;
 
@@ -99,6 +102,11 @@ class CountryController extends Controller
 
     public function destroy($id){
 
+        $validation = $this->validateDestroy($id);
+
+        if($validation !== "ok"){
+            return view('error')->with('errors', $validation);
+        }
         try{
             Country::findOrFail($id)->delete();
             return redirect('/country');
@@ -109,6 +117,17 @@ class CountryController extends Controller
 
         }
 
+    }
+
+    private function validateDestroy($id){
+
+        $events = Event::where('id_countries',$id)->exists();
+        $teams = Team::where('id_countries',$id)->exists();
+        $league = LeagueCountry::where('id_countries', $id)->exists();
+
+        if($events || $teams || $league)
+            return 'Cannot destroy country, because it is related to an entity';
+        return 'ok';
     }
 
 }
